Using a Parcelable was stupid, revert
In my defense, it was pretty late when I first had the bright idea of
using a Parcelable, and I thought RepoLoader.getModuleGroup(String) was
doing something heavier… which is why I figured getting the ModuleGroup
once would be more efficient.
I was revising the code when I noticed it's just stupid and I'm wrong.
This is much more readable and easier to maintain.
Sorry. :)